(maint) rename sources folder to adapters
