+++ adapters/internet/domain_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/overmindtech/sdpcache"
 )
 
-func TestDomainSourceGet(t *testing.T) {
+func TestDomainAdapterGet(t *testing.T) {
 	t.Parallel()
 
 	src := &DomainAdapter{
