+++ adapters/network/ip_test.go
