+++ go.mod
@@ -10,7 +10,7 @@ require (
 	github.com/nats-io/jwt/v2 v2.7.0
 	github.com/nats-io/nkeys v0.4.7
 	github.com/openrdap/rdap v0.9.2-0.20240517203139-eb57b3a8dedd
-	github.com/overmindtech/discovery v0.28.2
+	github.com/overmindtech/discovery v0.29.1
 	github.com/overmindtech/sdp-go v0.95.0
 	github.com/overmindtech/sdpcache v1.6.4
 	github.com/sirupsen/logrus v1.9.3
